completed deletePost method
Refactored addPost and handleSubmit methods. handleSubmit adds the post to the database and addPost uses that information to create the post on the front end. This allows for the id for the posts to be the same on the front end and back end. This helps with deletion and updating methods.